Modifications to ALU, addition of Zero flag
Modified ALU to relegate the selection of Immediate, Register, or Program Counter values to the CU.
Changed the names of the operands to better reflect the ALU operations.
Added a Zero flag, used in the PC branch and jump instructions.

Testbench modified accordingly to the new changes. Assertations may sometimes failed despite output and behavior being the expected one.